lock no jekyll

+++ linux/06-jekyll/configure.sh
@@ -2,5 +2,10 @@
 
 echo " - Jekyll..."
 
+LOCK=/tmp/workstation.lock.update
+[ -f ${LOCK} ] && return
+
 apt-get install -y ruby ruby-dev > /dev/null
 gem install jekyll bundler sass  > /dev/null
+
+> ${LOCK}
